requirements.txt
cryptography>=41.0.0
liboqs>=0.9.0
streamlit>=1.28.0
pandas>=2.0.0
numpy<2.0
matplotlib